#17 Move instance storage creation to the instance module

+++ server/src/main/scala/de/frosner/broccoli/instances/InstanceConfiguration.scala
@@ -2,13 +2,13 @@ package de.frosner.broccoli.instances
 
 import com.typesafe.config.Config
 
-final case class InstanceConfiguration(pollingFrequency: Int, storageConfiguration: InstanceStorageConfiguration)
+final case class InstanceConfiguration(pollingFrequency: Long, storageConfiguration: InstanceStorageConfiguration)
 
 object InstanceConfiguration {
   protected val log = play.api.Logger(getClass)
 
   def fromConfig(config: Config): InstanceConfiguration = {
-    val pollingFrequency = config.getInt("pollingFrequency")
+    val pollingFrequency = config.getLong("pollingFrequency")
     if (pollingFrequency <= 0) {
       throw new IllegalArgumentException(
         s"Invalid polling frequency specified: $pollingFrequency. Needs to be a positive integer.")
